converted nextjs to typescript
