+++ web/interfaces/firebase.ts
@@ -2,5 +2,5 @@ export interface Firebase {
   ref: string;
   uid?: string;
   isAbsolute?: boolean;
-  data?: any;
+  data: any;
 }
