--- web/middlewares/runMiddleware.js
@@ -1,12 +1,0 @@
-const runMiddleware = (req, res, middleware) => {
-  return new Promise((resolve, reject) => {
-    middleware(req, res, (result) => {
-      if (result instanceof Error) {
-        return reject(result);
-      }
-      return resolve(result);
-    });
-  });
-};
-
-export default runMiddleware;
