+++ web/package.json
@@ -30,7 +30,11 @@
   },
   "devDependencies": {
     "@svgr/webpack": "5.5.0",
-    "@types/speakeasy": "^2.0.6",
+    "@types/compression": "1.7.1",
+    "@types/cors": "2.8.11",
+    "@types/express": "4.17.13",
+    "@types/express-rate-limit": "5.1.3",
+    "@types/speakeasy": "2.0.6",
     "@typescript-eslint/eslint-plugin": "4.28.2",
     "eslint": "7.30.0",
     "eslint-config-next": "11.0.1",
